requirement.txt
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.1
cryptography==3.3.1
decorator==4.4.2
distro==1.5.0
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
future==0.18.2
google-api-core==1.24.1
google-auth==1.24.0
google-cloud==0.34.0
google-cloud-speech==2.0.1
google-cloud-texttospeech==2.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
isodate==0.6.0
libcst==0.3.16
lxml==4.6.2
mypy-extensions==0.4.3
networkx==2.5
nibabel==3.2.1
nipype==1.6.0
numpy==1.19.3
packaging==20.8
pandas==1.2.0
pathlib==1.0.1
pdfminer.six==20201018
Pillow==8.1.0
playsound==1.2.2
proto-plus==1.13.0
protobuf==3.14.0
prov==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
python-dateutil==2.8.1
pytz==2020.5
pyxnat==1.4
PyYAML==5.3.1
rdflib==5.0.0
requests==2.25.1
rsa==4.6
scipy==1.6.0
simplejson==3.17.2
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.1
tabula-py==2.2.0
tika==1.24
traits==6.1.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
XlsxWriter==1.3.7
